+++ examples/resources/artifactory_local_hex_repository/import.sh
@@ -1,0 +1,2 @@
+terraform import artifactory_local_hex_repository.my-hex-local my-hex-local
+