+++ examples/resources/artifactory_virtual_hex_repository/import.sh
@@ -1,0 +1,2 @@
+terraform import artifactory_virtual_hex_repository.my-hex-virtual my-hex-virtual
+