+++ pkg/artifactory/resource/repository/local/resource_artifactory_local_hex_repository.go
@@ -1,0 +1,121 @@
+package local
+
+import (
+	"context"
+	"reflect"
+
+	"github.com/hashicorp/terraform-plugin-framework/diag"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/jfrog/terraform-provider-artifactory/v12/pkg/artifactory/resource/repository"
+	"github.com/samber/lo"
+)
+
+func NewHexLocalRepositoryResource() resource.Resource {
+	return &localHexResource{
+		localResource: NewLocalRepositoryResource(
+			repository.HexPackageType,
+			repository.PackageNameLookup[repository.HexPackageType],
+			reflect.TypeFor[LocalHexResourceModel](),
+			reflect.TypeFor[LocalHexAPIModel](),
+		),
+	}
+}
+
+type localHexResource struct {
+	localResource
+}
+
+type LocalHexResourceModel struct {
+	LocalResourceModel
+	HexPrimaryKeyPairRef types.String `tfsdk:"hex_primary_keypair_ref"`
+}
+
+func (r *LocalHexResourceModel) GetCreateResourcePlanData(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	// Read Terraform plan data into the model
+	resp.Diagnostics.Append(req.Plan.Get(ctx, r)...)
+}
+
+func (r LocalHexResourceModel) SetCreateResourceStateData(ctx context.Context, resp *resource.CreateResponse) {
+	// Save data into Terraform state
+	resp.Diagnostics.Append(resp.State.Set(ctx, &r)...)
+}
+
+func (r *LocalHexResourceModel) GetReadResourceStateData(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	// Read Terraform state data into the model
+	resp.Diagnostics.Append(req.State.Get(ctx, r)...)
+}
+
+func (r LocalHexResourceModel) SetReadResourceStateData(ctx context.Context, resp *resource.ReadResponse) {
+	// Save data into Terraform state
+	resp.Diagnostics.Append(resp.State.Set(ctx, &r)...)
+}
+
+func (r *LocalHexResourceModel) GetUpdateResourcePlanData(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	// Read Terraform plan data into the model
+	resp.Diagnostics.Append(req.Plan.Get(ctx, r)...)
+}
+
+func (r *LocalHexResourceModel) GetUpdateResourceStateData(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	// Read Terraform state data into the model
+	resp.Diagnostics.Append(req.State.Get(ctx, r)...)
+}
+
+func (r LocalHexResourceModel) SetUpdateResourceStateData(ctx context.Context, resp *resource.UpdateResponse) {
+	// Save data into Terraform state
+	resp.Diagnostics.Append(resp.State.Set(ctx, &r)...)
+}
+
+func (r LocalHexResourceModel) ToAPIModel(ctx context.Context, packageType string) (interface{}, diag.Diagnostics) {
+	diags := diag.Diagnostics{}
+
+	model, d := r.LocalResourceModel.ToAPIModel(ctx, packageType)
+	if d != nil {
+		diags.Append(d...)
+	}
+
+	localAPIModel := model.(LocalAPIModel)
+	localAPIModel.RepoLayoutRef = r.RepoLayoutRef.ValueString()
+
+	return LocalHexAPIModel{
+		LocalAPIModel:        localAPIModel,
+		HexPrimaryKeyPairRef: r.HexPrimaryKeyPairRef.ValueString(),
+	}, diags
+}
+
+func (r *LocalHexResourceModel) FromAPIModel(ctx context.Context, apiModel interface{}) diag.Diagnostics {
+	diags := diag.Diagnostics{}
+
+	model := apiModel.(*LocalHexAPIModel)
+
+	r.LocalResourceModel.FromAPIModel(ctx, model.LocalAPIModel)
+	r.RepoLayoutRef = types.StringValue(model.RepoLayoutRef)
+	r.HexPrimaryKeyPairRef = types.StringValue(model.HexPrimaryKeyPairRef)
+
+	return diags
+}
+
+type LocalHexAPIModel struct {
+	LocalAPIModel
+	HexPrimaryKeyPairRef string `json:"primaryKeyPairRef"`
+}
+
+func (r *localHexResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+	attributes := lo.Assign(
+		LocalAttributes,
+		repository.RepoLayoutRefAttribute(r.Rclass, r.PackageType),
+		map[string]schema.Attribute{
+			"hex_primary_keypair_ref": schema.StringAttribute{
+				Required:            true,
+				MarkdownDescription: "Select the RSA key pair to sign and encrypt content for secure communication between Artifactory and the Mix client.",
+			},
+		},
+	)
+
+	resp.Schema = schema.Schema{
+		Version:     CurrentSchemaVersion,
+		Attributes:  attributes,
+		Description: r.Description,
+	}
+}