+++ CHANGELOG.md
@@ -4,6 +4,22 @@ BUG FIXES:
 
 * resource/artifactory_build_webhook, resource/artifactory_release_bundle_webhook, resource/artifactory_release_bundle_v2_webhook: Fix to send selected_builds, registered_release_bundle_names, and selected_release_bundles as null instead of empty array when any_build or any_release_bundle is set to true. PR: [#1333](https://github.com/jfrog/terraform-provider-artifactory/pull/1333)
 
+FEATURES:
+
+**New Resource:** `artifactory_local_hex_repository` to support local Hex repository. Issue: [#1230](https://github.com/jfrog/terraform-provider-artifactory/issues/1230) PR: [#1336](https://github.com/jfrog/terraform-provider-artifactory/pull/1336)
+
+**New Resource:** `artifactory_remote_hex_repository` to support remote Hex repository. Issue: [#1230](https://github.com/jfrog/terraform-provider-artifactory/issues/1230) PR: [#1336](https://github.com/jfrog/terraform-provider-artifactory/pull/1336)
+
+**New Resource:** `artifactory_virtual_hex_repository` to support virtual Hex repository. Issue: [#1230](https://github.com/jfrog/terraform-provider-artifactory/issues/1230) PR: [#1336](https://github.com/jfrog/terraform-provider-artifactory/pull/1336)
+
+IMPROVEMENTS:
+
+* data/artifactory_local_hex_repository: Migrated from Terraform Plugin SDKv2 to Terraform Plugin Framework. Issue: [#1230](https://github.com/jfrog/terraform-provider-artifactory/issues/1230) PR: [#1336](https://github.com/jfrog/terraform-provider-artifactory/pull/1336)
+
+* data/artifactory_remote_hex_repository: Migrated from Terraform Plugin SDKv2 to Terraform Plugin Framework. Issue: [#1230](https://github.com/jfrog/terraform-provider-artifactory/issues/1230) PR: [#1336](https://github.com/jfrog/terraform-provider-artifactory/pull/1336)
+
+* data/artifactory_virtual_hex_repository: Migrated from Terraform Plugin SDKv2 to Terraform Plugin Framework. Issue: [#1230](https://github.com/jfrog/terraform-provider-artifactory/issues/1230) PR: [#1336](https://github.com/jfrog/terraform-provider-artifactory/pull/1336)
+
 ### 12.10.1 (Oct 3, 2025). Tested on Artifactory 7.117.17 with Terraform 1.13.3 and OpenTofu 1.10.6
 
 BUG FIXES: